% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gz.R
\name{filter_gz}
\alias{filter_gz}
\title{Multiple knockoff filter by Gimenez & Zou (2018)}
\usage{
filter_gz(kappa, tau, fdr = 0.2, n_knockoff)
}
\arguments{
\item{kappa}{A \code{p} vector of test statistics, with kappa_i = 1 indicating the original variable winning}

\item{tau}{A \code{p} vector of test statistics, showing the manitude/importance of the variable}

\item{fdr}{The pre-specified upper bound level of controlled FDR.}

\item{n_knockoff}{number of knockoffs constructed}
}
\value{
The selected variable and their corresponding scores
}
\description{
This function implements the stable m-knockoff filter by Gimenez & Zou (2018)
given the test statistics (kappa, tau)
}
\examples{
library(mknockoff)
set.seed(123)
n <- 100
p <- 30
x <- matrix(data = rnorm(n * p), nrow = n, ncol = p)
y <- x[, 1] - 2 * x[, 2] + rnorm(n)
omega <- c(2, 9, sample(seq(2, 9), size = 28, replace = TRUE))
# construct multiple knockoffs
X_k <- knockoff_Gaussian(X = x, mu = rep(0, p), Sigma = diag(1, p), omega = omega)
# compute knockoff statistics
stat <- stat_glmnet_coef(X = x, X_k = X_k, y = y, omega = omega)
# run gz filter
gz_output <- filter_gz(kappa = stat$kappa, tau = stat$tau, fdr = 0.2, n_knockoff = max(omega))
}
