% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{generate_path}
\alias{generate_path}
\title{Multiple knockoff path}
\usage{
generate_path(kappa, tau)
}
\arguments{
\item{kappa}{A \code{p} vector of test statistics, with kappa_i = 1 indicating the original variable winning}

\item{tau}{A \code{p} vector of test statistics, showing the manitude/importance of the variable}
}
\value{
An list of selected variable sets
}
\description{
This function generates a path of selected variables using multiple knockoff
given the test statistics (kappa, tau)
}
\examples{
library(cheapknockoff)
set.seed(123)
n <- 100
p <- 30
x <- matrix(data = rnorm(n * p), nrow = n, ncol = p)
y <- x[, 1] - 2 * x[, 2] + rnorm(n)
omega <- c(2, 9, sample(seq(2, 9), size = 28, replace = TRUE))
# construct multiple knockoffs
X_k <- multiple_knockoff_Gaussian(X = x, mu = rep(0, p), Sigma = diag(1, p), omega = omega)
# compute knockoff statistics
stat <- cheapknockoff::stat_glmnet_coef(X = x, X_k = X_k, y = y, omega = omega)
# yield the path of selected variables
path <- cheapknockoff::generate_path(kappa = stat$kappa, tau = stat$tau)
}
