% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_X.R
\name{multiple_knockoff_Gaussian}
\alias{multiple_knockoff_Gaussian}
\title{Generating multiple knockoff variables}
\usage{
multiple_knockoff_Gaussian(X, mu, Sigma, omega, type = c("entropy", "sdp",
  "equi"), diag_s = NULL)
}
\arguments{
\item{X}{A \code{n}-by-\code{p} matrix of original variables}

\item{mu}{A \code{p} vector of mean parameter of the Gaussian distribution of the original variables}

\item{Sigma}{A \code{p}-by-\code{p} covariance matrix of the original variables}

\item{omega}{A \code{p} vector indicating the weights for each variable. For now, we require each entry to be an integer greater than or equal to 2.}

\item{type}{Could be "entropy" (default), "sdp" or "equal", indicating the method that will be used to construct the knockoff variables.}

\item{diag_s}{A \code{p} vector of pre-computed (user-specified) covariance between the original and the knockoff variables. This will be computed according to method, if not supplied.}
}
\value{
a \code{n}-by-\code{p * max{omega - 1}} matrix containing the constructed knockoff variables. Although we construct the same number of knockoffs for all variables (which is the maximum cost), in subsequent steps, we only use the number of knockoffs based on the feature costs.
}
\description{
This function samples multivariate Gaussian model-X knockoff variables multiple times for each original variable
}
\examples{
library(cheapknockoff)
set.seed(123)
n <- 100
p <- 30
x <- matrix(data = rnorm(n * p), nrow = n, ncol = p)
y <- x[, 1] - 2 * x[, 2] + rnorm(n)
omega <- c(2, 9, sample(seq(2, 9), size = 28, replace = TRUE))
# construct multiple knockoffs
X_k <- multiple_knockoff_Gaussian(X = x, mu = rep(0, p), Sigma = diag(1, p), omega = omega)
}
